LayoutTests/fast/forms/datetimelocal/datetimelocal-editable-components/datetimelocal-editable-components-mouse-events-expected.txt

Test for mouse events for <input type=datetime-local>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Enabled Input

PASS input.value is "2020-09-20T05:30"
PASS input.value is "2020-09-12T05:30"
PASS input.value is "3030-09-12T05:30"
PASS input.value is "3030-09-12T08:30"
PASS input.value is "3030-09-12T08:45"
PASS input.value is "3030-09-12T20:45"
PASS input.value is "3030-06-12T20:45"
PASS input.value is "3030-06-12T20:45"
PASS clickEventsFired is 7

Disabled Input

PASS clickEventsFired is 0

Readonly Input

PASS clickEventsFired is 7
PASS successfullyParsed is true

TEST COMPLETE
